Moves billing logic to a common place
