+++ android/java/org/chromium/chrome/browser/billing/PurchaseModel.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-package org.chromium.chrome.browser.vpn.billing;
+package org.chromium.chrome.browser.billing;
 
 import com.android.billingclient.api.Purchase;
