+++ android/java/org/chromium/chrome/browser/billing/sources.gni
@@ -3,12 +3,12 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at https://mozilla.org/MPL/2.0/.
 
-brave_app_vpn_billing_sources = [
-  "../../brave/android/java/org/chromium/chrome/browser/vpn/billing/InAppPurchaseWrapper.java",
-  "../../brave/android/java/org/chromium/chrome/browser/vpn/billing/PurchaseModel.java",
+brave_app_billing_sources = [
+  "//brave/android/java/org/chromium/chrome/browser/billing/InAppPurchaseWrapper.java",
+  "//brave/android/java/org/chromium/chrome/browser/billing/PurchaseModel.java",
 ]
 
-brave_app_vpn_billing_deps =
+brave_app_billing_deps =
     [ "//brave/third_party/android_deps:com_android_billingclient_java" ]
 
 brave_java_base_module_billing_deps = [
